refactor(desktop): extract release-feed interpretation into updater/feed
Mechanical move out of updater.ts: metadata fetch/parse, per-channel
version resolution, artifact selection with the payload-to-installer
fallback, checksum resolution, candidate/release naming and matching,
the installed-outer version read, and the installer-reinstall floor move
to updater/feed.ts; numberField/objectField join updater/support.
updater.ts re-exports compareVersions, resolveInstalledOuterVersion and
remoteRequiresReinstall so existing import sites are unchanged. The two
pure feed test cases move to tests/main/updater/feed.test.ts; total case
count is unchanged.
